convert tsne to float

+++ core/libs/facesengine/multicore_tsne/dimension_reducer.cpp
@@ -7,11 +7,11 @@ namespace Digikam
 
 cv::Mat DimensionReducer::project(cv::Mat data, int nbDimension) 
 {
-    cv::Mat projectedData(data.rows, nbDimension, CV_64F);
-    TSNE::tsne_run_double(reinterpret_cast<double*>(data.data), 
+    cv::Mat projectedData(data.rows, nbDimension, CV_32F);
+    TSNE::tsne_run_double(reinterpret_cast<float*>(data.data), 
                           data.rows, 
                           data.cols, 
-                          reinterpret_cast<double*>(projectedData.data), 
+                          reinterpret_cast<float*>(projectedData.data), 
                           nbDimension, 
                           30, 
                           0.5,
